Moved controller function for opportunity sections out to a new controller
