+++ src/main/scala/controllers/manage/OpportunityController.scala
@@ -3,19 +3,12 @@ package controllers.manage
 import javax.inject.Inject
 
 import actions.OpportunityAction
-import cats.data.Validated._
-import controllers.FieldCheckHelpers.hinting
-import controllers._
-import forms._
-import forms.validation.DateTimeRangeValues
 import models._
-import org.joda.time.LocalDate
 import org.joda.time.format.DateTimeFormat
-import play.api.libs.json._
 import play.api.mvc.{Action, Controller}
 import services.{ApplicationFormOps, OpportunityOps}
 
-import scala.concurrent.{ExecutionContext, Future}
+import scala.concurrent.ExecutionContext
 
 case class OpportunityLibraryEntry(id: OpportunityId, title: String, status: String, structure: String)
 
@@ -77,64 +70,6 @@ class OpportunityController @Inject()(opportunities: OpportunityOps, appForms: A
     }
   }
 
-  val SECTION_FIELD_NAME = "section"
-  val sectionField = TextAreaField(None, SECTION_FIELD_NAME, 500)
-
-  def doEditSection(opp: Opportunity, sectionNum: Int, initial: JsObject, errs: Seq[forms.validation.FieldError] = Nil) = {
-    val hints = FieldCheckHelpers.hinting(initial, Map(SECTION_FIELD_NAME -> sectionField.check))
-    opp.description.find(_.sectionNumber == sectionNum) match {
-      case Some(section) =>
-        val q = Question(section.description.getOrElse(""), None, section.helpText)
-        Ok(views.html.manage.editOppSectionForm(sectionField, opp, section,
-          routes.OpportunityController.editSection(opp.id, sectionNum).url, Map(SECTION_FIELD_NAME -> q), initial, errs, hints))
-      case None => NotFound
-    }
-
-  }
-
-  def editSection(id: OpportunityId, sectionNum: Int) = OpportunityAction(id).async { request =>
-    appForms.byOpportunityId(id).map {
-      case Some(appForm) =>
-        request.opportunity.description.find(_.sectionNumber == sectionNum) match {
-          case Some(sect) if sect.sectionType == OppSectionType.Text =>
-            val answers = JsObject(Seq(SECTION_FIELD_NAME -> Json.toJson(sect.text)))
-            doEditSection(request.opportunity, sectionNum, answers)
-          case Some(sect) => Ok(views.html.manage.whatWeWillAskPreview(request.uri, request.opportunity, sectionNum, appForm))
-          case None => NotFound
-        }
-      case None => NotFound
-    }
-  }
-
-  def saveSection(id: OpportunityId, sectionNum: Int) = OpportunityAction(id).async(JsonForm.parser) { implicit request =>
-    (request.body.values \ SECTION_FIELD_NAME).toOption.map { fValue =>
-      sectionField.check(SECTION_FIELD_NAME, fValue) match {
-        case Nil =>
-          opportunities.saveDescriptionSectionText(id, sectionNum, Some(fValue.as[String])).map { _ =>
-            request.body.action match {
-              case Preview =>
-                Redirect(controllers.manage.routes.OpportunityController.previewOppSection(id, sectionNum))
-                  .flashing(PREVIEW_BACK_URL_FLASH ->
-                    controllers.manage.routes.OpportunityController.editSection(id, sectionNum).url)
-              case _ =>
-                Redirect(controllers.manage.routes.OpportunityController.showOverviewPage(id))
-            }
-          }
-        case errors => Future.successful(doEditSection(request.opportunity, sectionNum, request.body.values, errors))
-      }
-    }.getOrElse(Future.successful(BadRequest))
-  }
-
-  def viewSection(id: OpportunityId, sectionNum: Int) = OpportunityAction(id).async { request =>
-    appForms.byOpportunityId(id).map {
-      case Some(appForm) => request.opportunity.publishedAt match {
-        case Some(_) => Ok(views.html.manage.viewOppSection(request.opportunity, appForm, sectionNum, request.flash.get(PREVIEW_BACK_URL_FLASH)))
-        case None => Redirect(controllers.manage.routes.OpportunityController.editSection(id, sectionNum))
-      }
-      case None => NotFound
-    }
-  }
-
   def duplicate(id: OpportunityId) = Action.async {
     request =>
       opportunities.duplicate(id).map {
@@ -158,29 +93,4 @@ class OpportunityController @Inject()(opportunities: OpportunityOps, appForms: A
     request =>
       Ok(views.html.manage.guidance(backUrl))
   }
-
-  def previewOppSection(id: OpportunityId, sectionid: Int) = OpportunityAction(id) {
-    request =>
-      Ok(views.html.manage.previewOppSection(request.opportunity, sectionid, request.flash.get(PREVIEW_BACK_URL_FLASH)))
-  }
-}
-
-sealed trait CreateOpportunityChoice {
-  def name: String
-}
-
-object CreateOpportunityChoice {
-  def apply(s: Option[String]): Option[CreateOpportunityChoice] = s match {
-    case Some(NewOpportunityChoice.name) => Some(NewOpportunityChoice)
-    case Some(ReuseOpportunityChoice.name) => Some(ReuseOpportunityChoice)
-    case _ => None
-  }
-}
-
-case object NewOpportunityChoice extends CreateOpportunityChoice {
-  val name = "new"
-}
-
-case object ReuseOpportunityChoice extends CreateOpportunityChoice {
-  val name = "reuse"
 }
